change DiaryTabController name, add TodayPage()
